week-7/parser/requirements.txt
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
marshmallow==3.19.0
marshmallow-dataclass==8.5.12
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pyarrow==11.0.0
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
typing-inspect==0.8.0
typing_extensions==4.5.0
